% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_8601}
\alias{from_8601}
\title{This function is called from the to_unixtime() function.
This takes a data-time object as a parameter and returns
a cleaner string by stripping out unnecessary objects
from the timestamp.}
\usage{
from_8601(x)
}
\arguments{
\item{x}{your timestamp passed from to_unixtime()

Mind that the returned pattern has 'UTC' format.}
}
\value{
a pattern/format of the timestamp to be converted
to unix time.
}
\description{
This function is called from the to_unixtime() function.
This takes a data-time object as a parameter and returns
a cleaner string by stripping out unnecessary objects
from the timestamp.
}
