% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tickets.R
\name{get_tickets}
\alias{get_tickets}
\title{Get Zendesk Tickets}
\usage{
get_tickets(email_id, token, subdomain, start_time = 0)
}
\arguments{
\item{email_id}{your Zendesk emailid}

\item{token}{your Zendesk API token}

\item{subdomain}{your organization's name}

\item{start_time}{Date or Datetime object to get all
tickets modified after that date.}
}
\value{
a Data Frame containing all tickets after the
start time
}
\description{
This function takes your Email Id, authentication token,
sub-domain and start time as parameters and gets all the
tickets which have been updated on or after the start
time parameter. By default each page returns 1000 unique
tickets and an "after_cursor" value which stores a
pointer to the next page. After getting the first page
it uses the pointer to fetch the subsequent pages.
}
\details{
The start time parameter here should be in 'UTC' format as
Zendesk uses the 'UTC' time format to export results. For eg
"2020-07-30 20:00:00 EDT" will be "2020-08-01 00:00:00 UTC".
So the user must do proper adjustment to the time format to
convert the desired time from their time zone to UTC format.
A wrong time format wont throw any error but it will not import
accurate information.

Its not a good practice to write down these authentication
parameters in your code. There are various methods and
packages available that are more secure; this package
doesn't require you to use any one in particular.
}
\examples{
\dontrun{
allTickets <- getTickets(email_id, token, subdomain, start_time = 0)
}
}
\references{
\url{https://developer.zendesk.com/rest_api
/docs/support/incremental_export#start_time}
}
