% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users.R
\name{get_users}
\alias{get_users}
\title{Get Metrics on All Zendesk Users.}
\usage{
get_users(email_id, token, subdomain)
}
\arguments{
\item{email_id}{Zendesk Email Id (username).}

\item{token}{Zendesk API token.}

\item{subdomain}{Your organization's Zendesk sub-domain.}
}
\value{
Data Frame with user details
}
\description{
This function takes your Email Id, authentication token,
sub-domain and parse all the users in a list.
Since each iteration only returns 100 users at a time
you must run the loop until the "next_page" parameter
is equal to null.
}
\details{
Its not a good practice to write down these authentication
parameters in your code. There are various methods and
packages available that are more secure; this package
doesn't require you to use any one in particular.
}
\examples{
\dontrun{
users <- get_users(email_id, token, subdomain)
}
}
\references{
\url{https://developer.zendesk.com/rest_api
/docs/support/users}
}
