% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTicketMetrics.R
\name{getAllTicketMetrics}
\alias{getAllTicketMetrics}
\title{This function is used to import all the tickets metrics}
\usage{
getAllTicketMetrics(email_id, token, subdomain)
}
\arguments{
\item{email_id}{your Zendesk emailid}

\item{token}{your Zendesk API token}

\item{subdomain}{your organization's name}
}
\value{
a Data Frame of all ticket metrics for all tickets
}
\description{
This function takes your emailid,authentication token,
sub-domain and parse all the tickets and its corresponding
metrics in a list. Since each iteration only returns 100
tickets at a time. We have to run the loop till we find
"next_page" parameter equal to null.
}
\details{
Its not a good practice to write down these authentication
parameters in your code. There are various methods and
packages available that are more secure; this package
doesnt require you to use any one in particular.
}
\examples{
\dontrun{
allTicketMetrics <- getAllTicketMetrics(email_id,token,subdomain,
start_time=0)
}
}
\references{
\url{https://developer.zendesk.com/rest_api
/docs/support/incremental_export#start_time}
}
