% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTickets.R
\name{getTickets}
\alias{getTickets}
\title{This function is used to import all the tickets from a
given starting date.}
\usage{
getTickets(email_id, token, subdomain, start_time = 0)
}
\arguments{
\item{email_id}{your Zendesk emailid}

\item{token}{your Zendesk API token}

\item{subdomain}{your organization's name}

\item{start_time}{your Date object to import tickets}
}
\value{
a Data Frame containing all tickets after the
start time
}
\description{
This function takes your emailid,authentication token,
sub-domain and start time as parameters and gets all the
tickets which have been updated on or after the start
time parameter.By default each page returns 1000 unique
tickets. Along with each page , the JSON returns a parameter
"after_cursor" which stores a pointer to the next page.
After getting the first page we use this parameter as a
query list with the API to fetch the subsequent pages.
}
\details{
Its not a good practice to write down these authentication
parameters in your code. There are various methods and
packages available that are more secure; this package
doesnt require you to use any one in particular.
}
\examples{
\dontrun{
allTickets <- getTickets(email_id,token,subdomain,start_time=0)
}
}
\references{
\url{https://developer.zendesk.com/rest_api
/docs/support/incremental_export#start_time}
}
