% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunction.R
\name{to_unixtime}
\alias{to_unixtime}
\title{This is a general purpose helper function for this
package useful in converting a time stamp to unix time}
\usage{
to_unixtime(x)
}
\arguments{
\item{x}{your date time stamp}
}
\description{
This function is inspired/borrowed from a github page
by Neal Richardson who did a similar work to convert the
general date time format to unix time.
}
\details{
The unix time stamp is a way to track time as a running
total of seconds. This count starts at the Unix Epoch on
January 1st, 1970 at UTC. Therefore, the unix time stamp
is merely the number of seconds between a particular date
and the Unix Epoch.
}
