% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_8601}
\alias{from_8601}
\title{Clean Date Time Inputs and Convert to UTC}
\usage{
from_8601(x)
}
\arguments{
\item{x}{Start time value provided by user.}
}
\value{
A datetime vector with the user's start time
converted to UTC time zone.
}
\description{
This function is called from the to_unixtime() function.
It takes a data-time object as a parameter and returns
a cleaner string by stripping out unnecessary objects
from the timestamp.
}
\keyword{internal}
