% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ticketfields.R
\name{get_ticketfields}
\alias{get_ticketfields}
\title{Get Zendesk Ticket Fields}
\usage{
get_ticketfields(email_id, token, subdomain)
}
\arguments{
\item{email_id}{Zendesk Email Id (username).}

\item{token}{Zendesk API token.}

\item{subdomain}{Your organization's Zendesk sub-domain.}
}
\value{
a Data Frame containing all ticket fields
}
\description{
This function takes your Email Id, authentication token,
sub-domain  as parameters and gets all the system and
custom fields available for a zendesk ticket.
}
\details{
Its not a good practice to write down these authentication
parameters in your code. There are various methods and
packages available that are more secure; this package
doesn't require you to use any one in particular.
}
\examples{
\dontrun{
all_ticketfields <- get_ticketfields(email_id, token, subdomain)
}
}
\references{
\url{https://developer.zendesk.com/rest_api
/docs/support/ticket_fields}
}
