% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_unixtime}
\alias{to_unixtime}
\title{Convert a Date or Datetime Object to Unix Datetime}
\usage{
to_unixtime(x)
}
\arguments{
\item{x}{Date or datetime string or object.}
}
\value{
Integer vector with unix time stamp.
}
\description{
This is a general purpose helper function for this
package useful in converting a time stamp to unix time.
}
\details{
This function is inspired/borrowed from a github page
by Neal Richardson who did a similar work to convert the
general date time format to unix time.

The unix time stamp is a way to track time as a running
total of seconds. This count starts at the Unix Epoch on
January 1st, 1970 at UTC. Therefore, the unix time stamp
is merely the number of seconds between a particular date
and the Unix Epoch.
}
\keyword{internal}
